provide informative msg when using wrong Node ver
the cli will now provide a message stating that the
version of node you are running is too old for our
CLI code
